+++ knotx-example/knotx-mocks/src/main/java/com/cognifide/knotx/mocks/knot/MockKnotHandler.java
@@ -21,6 +21,7 @@ import com.cognifide.knotx.dataobjects.AdapterRequest;
 import com.cognifide.knotx.dataobjects.ClientResponse;
 import com.cognifide.knotx.dataobjects.KnotContext;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Optional;
@@ -87,7 +88,7 @@ public class MockKnotHandler implements Handler<Message<KnotContext>> {
   }
 
   private JsonObject mergeResponseValues(JsonObject result, Pair<String, Optional<Object>> value) {
-    return new JsonObject().put(value.getLeft(), value.getRight().get());
+    return new JsonObject().put(value.getLeft(), value.getRight().orElse(StringUtils.EMPTY));
   }
 
 }